pyproject.toml
[project]
name = "arkline-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.0.15",
    "fastapi[all]>=0.116.1",
    "groq>=0.30.0",
    "langchain>=0.3.26",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.27",
    "langchain-groq>=0.3.6",
    "langchain-huggingface>=0.3.1",
    "numpy>=2.2.6",
    "pdfplumber>=0.11.7",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
    "sentence-transformers>=5.0.0",
    "torch>=2.7.1",
    "transformers>=4.53.3",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
